pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "advance-agent"
version = "0.1.0"
description = "Advanced Agent with bionic memory, powered by OpenAI SDK"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["advance_agent*"]
